LPS-65221 SF - Renaming for consistency
